basic support for image resizing.

+++ src/main/kotlin/com/dc2f/ContentDef.kt
@@ -6,8 +6,11 @@ import com.fasterxml.jackson.annotation.JacksonInject
 import com.vladsch.flexmark.html.HtmlRenderer
 import com.vladsch.flexmark.parser.Parser
 import mu.KotlinLogging
+import net.coobird.thumbnailator.Thumbnails
+import net.coobird.thumbnailator.geometry.Positions
 import java.lang.annotation.Inherited
 import java.nio.file.*
+import javax.imageio.ImageIO
 
 private val logger = KotlinLogging.logger {}
 
@@ -52,17 +55,83 @@ open class FileAsset(val file: ContentPath, val fsPath: Path) {
     fun href(context: RenderContext<*>): String {
         val targetPath = context.rootPath.resolve(file.toString())
         Files.createDirectories(targetPath.parent)
-        Files.copy(fsPath, targetPath)
+        if (!Files.exists(targetPath)) {
+            Files.copy(fsPath, targetPath)
+        }
         return "/$file"
     }
 }
+
+enum class FillType {
+    Fit,
+    Cover,
+    Transform
+}
+
+//data class ImageSize(
+//    val width: Int, val height: Int
+//) {
+//    fun resize(targetWidth: Int, targetHeight: Int, fillType: FillType) =
+//        when(fillType) {
+//            FillType.Transform -> ImageSize(targetWidth, targetHeight)
+//            FillType.Cover -> {
+//                // 10x100 -> 20x100 ---- (0.1 vs 0.2)
+//                // 100x10 -> 100x20 ---- (10 vs 5)
+//                val ratio = width.toDouble() / height
+//                val targetRatio = targetWidth.toDouble() / targetHeight
+//                if (ratio > targetRatio) {
+//                    ImageSize(targetWidth, (targetWidth / ratio).toInt())
+//                } else {
+//                    ImageSize((targetWidth * ratio).toInt(), targetHeight)
+//                }
+//            }
+//            FillType.Fit -> {
+//                val ratio = width.toDouble() / height
+//                val targetRatio = targetWidth.toDouble() / targetHeight
+//                if (ratio < targetRatio) {
+//                    ImageSize(targetWidth, (targetWidth / ratio).toInt())
+//                } else {
+//                    ImageSize((targetWidth * ratio).toInt(), targetHeight)
+//                }
+//            }
+//        }
+//}
+
+class ResizedImage(
+    val href: String,
+    val width: Int,
+    val height: Int
+)
 
 open class ImageAsset(file: ContentPath, fsPath: Path) : FileAsset(file, fsPath) {
     val imageInfo: ImageInfo by lazy { parseImage() }
     val width get() = imageInfo.width
     val height get() = imageInfo.height
 
-    fun parseImage() =
+    fun resize(context: RenderContext<*>, width: Int, height: Int, fillType: FillType): ResizedImage {
+        val targetPathOrig = context.rootPath.resolve(file.toString())
+        val fileName = "${fillType}_${width}x${height}_${targetPathOrig.fileName}"
+        val targetContentPath = file.sibling(fileName)
+        val targetPath = targetPathOrig.resolveSibling(fileName)
+
+        val original = ImageIO.read(fsPath.toFile())
+//        val targetSize = ImageSize(original.width, original.height)
+//            .resize(width, height, fillType)
+        val thumbnails = Thumbnails.of(original)
+        when(fillType) {
+            FillType.Cover -> thumbnails.size(width, height).crop(Positions.CENTER)
+            FillType.Fit -> thumbnails.size(width, height)
+            FillType.Transform -> thumbnails.forceSize(width, height)
+        }
+        thumbnails.toFile(targetPath.toFile())
+//        thumbnails.addFilter()
+        return ResizedImage(
+            "/$targetContentPath",
+            width,
+            height)
+    }
+
+    private fun parseImage() =
         ImageUtil.readImageData(fsPath) ?: throw IllegalArgumentException("Invalid image at $fsPath")
 }
